+++ Bombom/src/main/java/com/kh/bom/common/page/ProAjaxPageBarFactory.java
@@ -2,7 +2,7 @@ package com.kh.bom.common.page;
 
 public class ProAjaxPageBarFactory {
 
-	public static String getAjaxPageBar(int totalData,int cPage,int numPerpage,String url,String category) {
+	public static String getAjaxPageBar(int totalData,int cPage,int numPerpage,String url,String sort,String category) {
 
 		String pageBar="";
 		
@@ -24,7 +24,7 @@ public class ProAjaxPageBarFactory {
 		}else {
 			//페이지번호가 1이 아닐경우에는 이전버튼 선택시 (현재페이지-1)한 페이지수로 이동.
 		
-			pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+(pageNo-1)+",\'"+category+"\')\"> << </a>";
+			pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+(pageNo-1)+",\'"+sort+"',\'"+category+"\')\"> << </a>";
 			
 		}
 		
@@ -39,7 +39,7 @@ public class ProAjaxPageBarFactory {
 			}else {
 				//현재페이지랑 다른 페이지넘버를 선택할경우 그 페이지넙버로 이동
 				
-				pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+pageNo+",\'"+category+"\')\">"+pageNo+"</a>";
+				pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+pageNo+",\'"+sort+"',\'"+category+"\')\">"+pageNo+"</a>";
 				
 			}
 			pageNo++;
@@ -54,7 +54,7 @@ public class ProAjaxPageBarFactory {
 		}else {
 			//아닐경우 다음페이지시 다음페이지로이동(위에서 pageNo++해야지만 넘어오기에 pageNo로 이동하면 됨)
 	
-			pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+pageNo+",\'"+category+"\')\"> >> </a>";
+			pageBar+="<a class='w3-button' href=\"javascript:fn_paging("+pageNo+",\'"+sort+"',\'"+category+"\')\"> >> </a>";
 		
 		}
 		
@@ -62,11 +62,12 @@ public class ProAjaxPageBarFactory {
 		
 		pageBar+="<script>";
 
-		pageBar+="function fn_paging(cPage,category){";
-				
+		pageBar+="function fn_paging(cPage,sort,category){";
+		pageBar+="console.log(category);";
+		pageBar+="console.log(sort);";
 		pageBar+="$.ajax({";
 		pageBar+="url:'"+url+"',";
-		pageBar+="data:{cPage:cPage,numPerpage:8,sort:category},";
+		pageBar+="data:{cPage:cPage,numPerpage:8,sort:sort,category:category},";
 		pageBar+="type:'get',";
 		pageBar+="success:data=>{";
 		pageBar+="console.log(data);";
