+++ Bombom/src/main/java/com/kh/bom/product/controller/ProductController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.kh.bom.common.page.AjaxPageBarFactory;
-import com.kh.bom.common.page.ProAjaxPageBarFactory2;
+import com.kh.bom.common.page.ProAjaxPageBarFactory;
 import com.kh.bom.common.page.ProPageBarFactory;
 import com.kh.bom.inquiry.model.vo.Inquiry;
 import com.kh.bom.member.model.vo.Member;
@@ -67,7 +67,7 @@ public class ProductController {
 			m.addObject("list",service.selectProductList(cPage,numPerpage,sort,category));
 		}
 		
-		m.addObject("pageBar",ProAjaxPageBarFactory2.getAjaxPageBar(count, cPage, numPerpage, "productListAjax",sort,category));
+		m.addObject("pageBar",ProAjaxPageBarFactory.getAjaxPageBar(count, cPage, numPerpage, "productListAjax",sort,category));
 		m.addObject("cPage",cPage);
 		m.addObject("category",category);
 		m.addObject("sort",sort);
